Created ManifestsIndex out of ManifestsCache.
The cache is meant to represent only partial info of what is stored in the local node.
The cache should give fast reads to the client, but we cannot rely on the cache
to get all the information we want. In fact, the cache has a limited size and that
should be taken into account.

The index contains information that can be derived by reading all the manifests stored in the local node
(except for the HEADS) and thus it can be recreated at any time.
The index increases reads on fetching info that would otherwise require lots of disk reading.
For example: how to get all the invariants of type VERSION? or how to get all the tips of an asset?